feat: implement dashboard layout and pages with authentication
