+++ apps/dashboard/middleware.ts
@@ -1,0 +1,1 @@
+export { auth as middleware } from './auth';
