+++ apps/dashboard/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "dependencies": {
     "next": "~15.2.4",
-    "next-auth": "^4.24.11",
+    "next-auth": "^5.0.0-beta.29",
     "next-intl": "^4.3.4",
     "react": "19.0.0",
     "react-dom": "19.0.0"
