+++ apps/dashboard/src/app/(auth)/login/page.tsx
@@ -1,0 +1,22 @@
+import SignIn from '@dash/components/signin';
+
+export default function PageLogin() {
+  return (
+    <>
+      <div className="flex min-h-full flex-1 flex-col justify-center px-6 py-12 lg:px-8">
+        <div className="sm:mx-auto sm:w-full sm:max-w-sm">
+          <img
+            alt="Your Company"
+            src="/images/logo.svg"
+            className="mx-auto h-10 w-auto"
+          />
+          <h2 className="mt-10 text-center text-2xl/9 font-bold tracking-tight text-gray-900">
+            Sign in to your account
+          </h2>
+        </div>
+
+        <SignIn />
+      </div>
+    </>
+  );
+}
