+++ apps/dashboard/src/common/helpers/index.ts
@@ -1,17 +1,1 @@
-import type {
-  GetServerSidePropsContext,
-  NextApiRequest,
-  NextApiResponse,
-} from 'next';
-
-import { getServerSession } from 'next-auth';
-import { authOptions } from '@dash/app/api/auth/[...nextauth]/route';
-
-export function auth(
-  ...args:
-    | [GetServerSidePropsContext['req'], GetServerSidePropsContext['res']]
-    | [NextApiRequest, NextApiResponse]
-    | []
-) {
-  return getServerSession(...args, authOptions);
-}
+export { auth } from '@auth';
