+++ apps/dashboard/src/types/next-auth.d.ts
@@ -1,43 +1,43 @@
-import NextAuth from 'next-auth';
+// import NextAuth from 'next-auth';
 
-declare module 'next-auth' {
-  /**
-   * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
-   */
-  interface Session {
-    user: {
-      id: string;
-      email: string;
-      /** The user's postal address. */
-      address: string;
-      accessToken?: string; // Optional access token
-      refreshToken?: string; // Optional refresh token
-      expiresIn?: number; // Optional expiration time for the access token
-    };
-  }
+// declare module 'next-auth' {
+//   /**
+//    * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
+//    */
+//   interface Session {
+//     user: {
+//       id: string;
+//       email: string;
+//       /** The user's postal address. */
+//       address: string;
+//       accessToken?: string; // Optional access token
+//       refreshToken?: string; // Optional refresh token
+//       expiresIn?: number; // Optional expiration time for the access token
+//     };
+//   }
 
-  interface User {
-    id: string;
-    email: string;
-    accessToken?: string; // Optional access token
-    refreshToken?: string; // Optional refresh token
-  }
+//   interface User {
+//     id: string;
+//     email: string;
+//     accessToken?: string; // Optional access token
+//     refreshToken?: string; // Optional refresh token
+//   }
 
-  interface JWT {
-    id: string;
-    email: string;
-    accessToken?: string; // Optional access token
-    expiresIn?: number;
-    refreshToken?: string; // Optional refresh token
-  }
-}
+//   interface JWT {
+//     id: string;
+//     email: string;
+//     accessToken?: string; // Optional access token
+//     expiresIn?: number;
+//     refreshToken?: string; // Optional refresh token
+//   }
+// }
 
-declare module 'next-auth/jwt' {
-  interface JWT {
-    id: string;
-    email: string;
-    accessToken?: string; // Optional access token
-    expiresIn?: number;
-    refreshToken?: string; // Optional refresh token
-  }
-}
+// declare module 'next-auth/jwt' {
+//   interface JWT {
+//     id: string;
+//     email: string;
+//     accessToken?: string; // Optional access token
+//     expiresIn?: number;
+//     refreshToken?: string; // Optional refresh token
+//   }
+// }
